备份9

+++ sources/TZ_MCF52259_PWM.h
@@ -3,7 +3,7 @@
 
 #define TPWM0_SetDTY(x)     MCF_PWM_PWMDTY0=x
 #define TPWM1_SetDTY(x)     MCF_PWM_PWMDTY1=x
-#define TPWM45_SetDTY(x)    MCF_PWM_PWMDTY4 = (uint8)(x>>8),MCF_PWM_PWMDTY5 = (uint8)(x&0xff)
+#define TPWM45_SetDTY(x)    MCF_PWM_PWMDTY4 = x>>8,MCF_PWM_PWMDTY5 = x&0xff
 
 extern uint16 pwmCnt;
 extern uint8 aa3;
